+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-07-12
+
+### Added
+- Agent Icons support using Nerd font or Font Awesome
+- User config file at `~/.config/aot/aot.conf`, also supporting env vars and CLI options
+- Include Copilot agent
+
+### Changed
+- Updated README and added installation instructions
+
 ## [0.1.4] - 2026-07-09
 
 ### Added